added highlight routes but still have errors in passing the selection object and the getRangeAt(0)
