+++ backend/routes/siteRouter.js
@@ -110,6 +110,7 @@ siteRouter.route('/updateReviews').post((req, res, next) => {
     }
   });
 });
+
 /**
  * Usage: Probably when the user begins the access a new page.
  * ex:
@@ -137,4 +138,67 @@ siteRouter.route('/getSiteData').get((req, res, next) => {
   });
 });
 
+
+/**
+ * Usage: Add's highlights to backend
+ Payload: {
+ * "_id":"{url [REQUIRED]}",
+ * "userID": {User's ID}
+ * "highlightselection":Highlight selection,
+ * POST: Add highlight to website
+ */
+  siteRouter.route('/addHighlights').post((req, res, next) => {
+    console.log("POST: Updating website's number of reviews:", req.body);
+    console.log("Hightlight Selection type:", typeof(req.body.highlightSelection))
+    VisitedSites.findOne({ _id: req.body.url}, (err, results) => {
+      if (err || results == null) {
+        return res
+          .status(400)
+          .send({ message: 'Error in finding website in DB' });
+      } else {
+        results.highlights.push({"owner": req.body.userID,
+      "selection": req.body.highlightSelection})
+        results.save((error) => {
+          if (error) {
+            return res
+              .status(400)
+              .send({ message: 'Error occured in adding highlight' });
+          } else {
+            return res
+              .status(200)
+              .send({ message: "Updated website's stored highlights" });
+          }
+        });
+      }
+    });
+  });
+
+  /**
+ * Usage: Probably when the user begins the access a new page.
+ * ex:
+ * send with a payload that at least has
+ *
+ * Query params
+ *  "url":"{currentURL from user}",
+ *  "userID": {userID}
+ * GET: Get user's saved higlights on site
+ */
+siteRouter.route('/getUserHighlights').get((req, res, next) => {
+  if (req.query.url == null) {
+    console.log('ERROR: Null query received');
+    return res.status(400).send({ message: 'Need a valid site URL' });
+  }
+  console.log('GET: Information about current site:', req.query.url);
+  VisitedSites.findOne({ _id: req.query.url }, (err, result) => {
+    if (err || result == null) {
+      console.log(result);
+      console.log('Error:', err);
+      return res.status(400).send({ message: 'Current site not found' });
+    } else {
+      console.log('Result:', result);
+      let highlights = result.highlights.filter(highlight => highlight.owner == req.query.userID);
+      return res.send(highlights);
+    }
+  });
+});
 module.exports = siteRouter;
