+++ src/pages/Content/modules/StarRating-Component.js
